@@ -31,56 +31,48 @@
     "@golevelup/ts-jest": "^0.3.5",
     "@nestjs/schematics": "10.0.2",
     "@nestjs/testing": "10.2.4",
-    "@ngneat/spectator": "^15.0.0",
-    "@nx/angular": "17.1.3",
-    "@nx/devkit": "17.1.3",
-    "@nx/eslint": "17.1.3",
-    "@nx/eslint-plugin": "17.1.3",
-    "@nx/jest": "17.1.3",
-    "@nx/js": "17.1.3",
-    "@nx/nest": "17.1.3",
-    "@nx/node": "17.1.3",
-    "@nx/playwright": "17.1.3",
-    "@nx/web": "17.1.3",
-    "@nx/webpack": "17.1.3",
-    "@nx/workspace": "17.1.3",
+    "@ngneat/spectator": "^16.0.0",
+    "@nx/angular": "17.2.8",
+    "@nx/devkit": "17.2.8",
+    "@nx/eslint": "17.2.8",
+    "@nx/eslint-plugin": "17.2.8",
+    "@nx/jest": "17.2.8",
+    "@nx/js": "17.2.8",
+    "@nx/nest": "17.2.8",
+    "@nx/node": "17.2.8",
+    "@nx/playwright": "17.2.8",
+    "@nx/web": "17.2.8",
+    "@nx/webpack": "17.2.8",
+    "@nx/workspace": "17.2.8",
     "@playwright/test": "^1.39.0",
     "@schematics/angular": "17.0.0",
-    "@swc-node/register": "~1.6.7",
-    "@swc/cli": "~0.1.62",
-    "@swc/core": "~1.3.85",
-    "@swc/jest": "0.2.20",
     "@total-typescript/ts-reset": "^0.3.7",
     "@trivago/prettier-plugin-sort-imports": "^4.1.1",
     "@types/cron": "^2.0.1",
+    "@types/dompurify": "^3.0.5",
     "@types/jest": "29.4.4",
-    "@types/node": "20.10.2",
-    "@typescript-eslint/eslint-plugin": "6.10.0",
-    "@typescript-eslint/parser": "6.10.0",
-    "autoprefixer": "^10.4.0",
+    "@types/node": "20.11.17",
+    "@types/trusted-types": "^2.0.7",
+    "@typescript-eslint/eslint-plugin": "6.14.0",
+    "@typescript-eslint/parser": "6.14.0",
     "cobertura-merge": "^1.0.4",
-    "eslint": "8.55.0",
-    "eslint-config-prettier": "9.0.0",
+    "eslint": "8.56.0",
+    "eslint-config-prettier": "9.1.0",
     "eslint-plugin-jest": "^27.2.1",
-    "eslint-plugin-playwright": "^0.19.0",
+    "eslint-plugin-playwright": "^0.22.0",
     "eslint-plugin-rxjs": "^5.0.2",
-    "husky": "^8.0.0",
+    "husky": "^9.0.0",
     "jest": "29.5.0",
     "jest-environment-jsdom": "29.5.0",
     "jest-environment-node": "^29.5.0",
     "jest-preset-angular": "13.1.4",
     "lcov-result-merger": "^5.0.0",
     "mongo-migrate-ts": "^1.1.4",
-    "nx": "17.1.3",
+    "nx": "17.2.8",
     "playwright": "^1.39.0",
-    "postcss": "^8.4.31",
-    "prettier": "3.1.0",
-    "prettier-plugin-tailwindcss": "^0.5.0",
-    "supertest": "^6.3.3",
-    "swc-loader": "0.2.3",
-    "tailwindcss": "^3.0.2",
+    "prettier": "3.2.5",
     "ts-jest": "29.1.0",
-    "ts-node": "10.9.1",
+    "ts-node": "10.9.2",
     "typescript": "5.2.2",
     "wait-on": "^7.0.1"
   },
