@@ -3,7 +3,7 @@ const { resolve } = require('path');
 
 module.exports = composePlugins(withNx(), (config) => {
 	config.entry['workers/nina-warnings.worker'] = resolve(
-		'libs/api/warnings/src/lib/infra/worker/nina-warnings.worker.ts',
+		'libs/api/warning/src/lib/infra/worker/nina-warnings.worker.ts',
 	);
 
 	return config;
