@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 
-import { GEOSPATIAL_SERVICE } from './core/geospatial.service';
-import { GeospatialServiceImpl } from './infra/geospatial.service';
 import { GEOSPATIAL_TYPE_CONVERTER } from './core/geospatial-types.converter';
+import { GEOSPATIAL_SERVICE } from './core/geospatial.service';
 import { GeospatialTypesConverterImpl } from './infra/geospatial-types.converter';
+import { GeospatialServiceImpl } from './infra/geospatial.service';
 
 @Module({
 	providers: [
