@@ -1,22 +1,25 @@
+import { DeepMocked, createMock } from '@golevelup/ts-jest';
+import { Logger } from '@nestjs/common';
 import * as Sentry from '@sentry/node';
 
 import { PresentableException } from '@kordis/api/shared';
 
 import { SentryExceptionsFilter } from './sentry-exceptions.filter';
 
-describe('ExceptionsFilter', () => {
+describe('SentryExceptionsFilter', () => {
 	let sentryExceptionsFilter: SentryExceptionsFilter;
 	let addBreadcrumbMock: jest.Mock;
 	let captureExceptionMock: jest.Mock;
-
-	beforeEach(() => {
+	let logger: DeepMocked<Logger>;
+	beforeEach(async () => {
 		addBreadcrumbMock = jest.fn();
 		captureExceptionMock = jest.fn();
+		logger = createMock<Logger>();
 
 		(Sentry.addBreadcrumb as jest.Mock) = addBreadcrumbMock;
 		(Sentry.captureException as jest.Mock) = captureExceptionMock;
 
-		sentryExceptionsFilter = new SentryExceptionsFilter();
+		sentryExceptionsFilter = new SentryExceptionsFilter(logger);
 	});
 
 	afterEach(() => {
@@ -36,15 +39,11 @@ describe('ExceptionsFilter', () => {
 
 		sentryExceptionsFilter.catch(presentableException);
 
-		expect(addBreadcrumbMock).toHaveBeenCalledTimes(1);
-		expect(addBreadcrumbMock).toHaveBeenCalledWith({
-			level: 'error',
-			message: 'message',
-			data: {
-				name: 'Error',
-				code: 'code',
-				stack: expect.any(String),
-			},
+		expect(logger.log).toHaveBeenCalledTimes(1);
+		expect(logger.log).toHaveBeenCalledWith('message', {
+			name: 'Error',
+			code: 'code',
+			stack: expect.any(String),
 		});
 		expect(captureExceptionMock).not.toHaveBeenCalled();
 	});
@@ -58,6 +57,14 @@ describe('ExceptionsFilter', () => {
 		expect(captureExceptionMock).toHaveBeenCalledWith(exception, {
 			level: 'error',
 		});
+		expect(logger.error).toHaveBeenCalledTimes(1);
+		expect(logger.error).toHaveBeenCalledWith(
+			'Caught unhandled exception that was not presentable',
+			undefined,
+			{
+				exception,
+			},
+		);
 		expect(addBreadcrumbMock).not.toHaveBeenCalled();
 	});
 });
