@@ -4,30 +4,38 @@ import {
 	Injectable,
 	NestInterceptor,
 } from '@nestjs/common';
-import { GqlExecutionContext } from '@nestjs/graphql';
+import { GqlContextType, GqlExecutionContext } from '@nestjs/graphql';
 import { trace } from '@opentelemetry/api';
 import * as Sentry from '@sentry/node';
 import { Observable } from 'rxjs';
 
-import { KordisGqlContext } from '@kordis/api/shared';
+import { KordisGqlContext, KordisRequest } from '@kordis/api/shared';
+import { AuthUser } from '@kordis/shared/auth';
 
 @Injectable()
 export class SentryOTelUserContextInterceptor implements NestInterceptor {
 	intercept(context: ExecutionContext, next: CallHandler): Observable<unknown> {
-		const ctx = GqlExecutionContext.create(context);
-		const { user } = ctx.getContext<KordisGqlContext>().req;
+		let user: AuthUser | undefined; // user can be undefined, since we have routes passed through the auth interceptor which don't require a user (/health-check)
+		if (context.getType<GqlContextType>() === 'graphql') {
+			const ctx = GqlExecutionContext.create(context);
+			user = ctx.getContext<KordisGqlContext>().req.user;
+		} else {
+			user = context.switchToHttp().getRequest<KordisRequest>().user;
+		}
 
-		trace.getActiveSpan()?.setAttributes({
-			'user.id': user.id,
-			'user.email': user.email,
-			'user.name': `${user.firstName} ${user.lastName}`,
-		});
+		if (user) {
+			trace.getActiveSpan()?.setAttributes({
+				'user.id': user.id,
+				'user.email': user.email,
+				'user.name': `${user.firstName} ${user.lastName}`,
+			});
 
-		Sentry.setUser({
-			id: user.id,
-			email: user.email,
-			username: `${user.firstName} ${user.lastName}`,
-		});
+			Sentry.setUser({
+				id: user.id,
+				email: user.email,
+				username: `${user.firstName} ${user.lastName}`,
+			});
+		}
 
 		return next.handle();
 	}
