@@ -7,8 +7,8 @@ import {
 	ValidateNested,
 } from 'class-validator';
 
-import { BaseEntityModel } from '@kordis/api/shared';
 import { BBox, Coordinate, IsBBox } from '@kordis/api/geospatial';
+import { BaseEntityModel } from '@kordis/api/shared';
 
 @ObjectType()
 @InputType('OrganizationGeoSettingsInput')
